Show hex for +/-{inf|subnorm|nan} floats in test-emulate.
- Change test-emulate --log-operands to show hex output for
  floating point numbers that are classified +/-{inf|subnorm|nan}
- Add 'dyn' round mode enum for 0b111 (dynamic rounding mode)
- Add 'dyn' symbol to disassembler round mode output. The
  specification states that 0b111 selects dynamic rounding mode.
- Rename disassembler unknown round mode from 'unk' to 'inv'
  as a mnemonic for Invalid. The specification states Invalid.
- Fix return type of f32_classify and f64_classify.

+++ src/emu/riscv-fpu.h
@@ -44,7 +44,7 @@ namespace riscv {
 	inline float f32_sqrt(float a) { return std::sqrt(a); }
 	inline double f64_sqrt(double a) { return std::sqrt(a); }
 
-	inline float f32_classify(float a)
+	inline int f32_classify(float a)
 	{
 		union { uint32_t wu; float f; } v = { .f = a };
 		uint32_t neg = v.wu >> 31, exp = (v.wu >> 23) & ((1<<8)-1), man = v.wu & ((1<<23)-1);
@@ -61,7 +61,7 @@ namespace riscv {
 		}
 	}
 
-	inline double f64_classify(double a)
+	inline int f64_classify(double a)
 	{
 		union { uint64_t lu; double d; } v = { .d = a };
 		uint64_t neg = v.lu >> 63, exp = (v.lu >> 52) & ((1ULL<<11)-1), man = v.lu & ((1ULL<<52)-1);
